fixed variant.h (use PIN_Ax instead of Ax) to reenable sanity checks
In the arduino sources PIN_Ax are defined as numbers, which then gets
assigned to Ax.

For example:
  #define PIN_A7   (61)
  static const uint8_t A7 = PIN_A7;

'#if x == y' as used in the sanity checks seems not to work with the
const uint8_t values. Therefore in variant.h all occurences of Ax have
been replaced by PIN_Ax.